+++ test/distance_test.cc
@@ -20,6 +20,8 @@
 
 using namespace improbable::phtree;
 
+// NOTE: These are very superficial tests. Proper testing is done in the respective PhTree tests.
+
 TEST(PhTreeDistanceTest, DoubleEuclidean) {
     auto distance = DistanceEuclidean<2>();
     ASSERT_DOUBLE_EQ(5, distance(PhPointD<2>{-1, -1}, PhPointD<2>{2, 3}));
@@ -28,6 +30,26 @@ TEST(PhTreeDistanceTest, DoubleEuclidean) {
 TEST(PhTreeDistanceTest, DoubleL1) {
     auto distance = DistanceL1<2>();
     ASSERT_DOUBLE_EQ(7, distance(PhPointD<2>{-1, -1}, PhPointD<2>{2, 3}));
+}
+
+TEST(PhTreeDistanceTest, DoubleChebyshev) {
+    auto distance = DistanceChebyshev<2>();
+    ASSERT_DOUBLE_EQ(4, distance(PhPointD<2>{-1, -1}, PhPointD<2>{2, 3}));
+}
+
+TEST(PhTreeDistanceTest, FloatEuclidean) {
+    auto distance = DistanceEuclidean<2>();
+    ASSERT_DOUBLE_EQ(5, distance(PhPointF<2>{-1, -1}, PhPointF<2>{2, 3}));
+}
+
+TEST(PhTreeDistanceTest, FloatL1) {
+    auto distance = DistanceL1<2>();
+    ASSERT_DOUBLE_EQ(7, distance(PhPointF<2>{-1, -1}, PhPointF<2>{2, 3}));
+}
+
+TEST(PhTreeDistanceTest, FloatChebyshev) {
+    auto distance = DistanceChebyshev<2>();
+    ASSERT_DOUBLE_EQ(4, distance(PhPointF<2>{-1, -1}, PhPointF<2>{2, 3}));
 }
 
 TEST(PhTreeDistanceTest, LongEuclidean) {
@@ -39,3 +61,9 @@ TEST(PhTreeDistanceTest, LongL1) {
     auto distance = DistanceL1<2>();
     ASSERT_DOUBLE_EQ(7, distance(PhPoint<2>{-1, -1}, PhPoint<2>{2, 3}));
 }
+
+TEST(PhTreeDistanceTest, LongChebyshev) {
+    auto distance = DistanceChebyshev<2>();
+    ASSERT_DOUBLE_EQ(4, distance(PhPoint<2>{-1, -1}, PhPoint<2>{2, 3}));
+}
+